[NVSHAS-6761] NV-5.1 Secrisk>Compliance "Refresh" resets AdvancedFilter
Persist Compliance advanced fiter through refresh action.
Fix bug with Compliance advanced filter service name field.
