+++ admin/webapp/websrc/app/routes/compliance/compliance.filter.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import { Compliance } from '@common/types';
 import { TranslateService } from '@ngx-translate/core';
 import { BehaviorSubject } from 'rxjs';
 
@@ -54,9 +55,118 @@ export class ComplianceFilterService {
     this.filteredSubject$.next(this.isAdvFilterOn() || this._filtered);
   }
 
-  resetFilter() {
-    this._advFilter = this.initAdvFilter();
-    this.filteredSubject$.next(false);
+  resetFilter(filter?) {
+    this._advFilter = filter ? filter : this.initAdvFilter();
+    this.filteredSubject$.next(!!filter);
+  }
+
+  filterFn(compliance) {
+    let result = true;
+    if (
+      !this.advFilter.category.custom ||
+      !this.advFilter.category.docker ||
+      !this.advFilter.category.kubernetes ||
+      !this.advFilter.category.image
+    ) {
+      if (!this.advFilter.category.docker)
+        result = result && compliance.category !== 'docker';
+      if (!this.advFilter.category.custom)
+        result = result && compliance.category !== 'custom';
+      if (!this.advFilter.category.kubernetes)
+        result = result && compliance.category !== 'kubernetes';
+      if (!this.advFilter.category.image)
+        result = result && compliance.category !== 'image';
+    }
+    if (
+      this.advFilter.tags.gdpr ||
+      this.advFilter.tags.hipaa ||
+      this.advFilter.tags.nist ||
+      this.advFilter.tags.pci
+    ) {
+      if (compliance.tags && compliance.tags.length > 0) {
+        if (this.advFilter.tags.gdpr)
+          result = result && compliance.tags.includes('GDPR');
+        if (this.advFilter.tags.hipaa)
+          result = result && compliance.tags.includes('HIPAA');
+        if (this.advFilter.tags.nist)
+          result = result && compliance.tags.includes('NIST');
+        if (this.advFilter.tags.pci)
+          result = result && compliance.tags.includes('PCI');
+      } else return false;
+    }
+    if (this.advFilter.scoredType !== 'all') {
+      result =
+        result && compliance.scored.toString() === this.advFilter.scoredType;
+    }
+    if (this.advFilter.profileType !== 'all') {
+      result = result && compliance.profile === this.advFilter.profileType;
+    }
+    if (this.advFilter.containerName) {
+      if (compliance.workloads.length) {
+        result = this.checkEntity(
+          this.advFilter.matchTypes['Container'].id,
+          compliance.workloads,
+          this.advFilter.containerName,
+          result
+        );
+      } else return false;
+    }
+    if (this.advFilter.nodeName) {
+      if (compliance.nodes.length) {
+        result = this.checkEntity(
+          this.advFilter.matchTypes['Node'].id,
+          compliance.nodes,
+          this.advFilter.nodeName,
+          result
+        );
+      } else return false;
+    }
+    if (this.advFilter.imageName) {
+      if (compliance.images.length) {
+        result = this.checkEntity(
+          this.advFilter.matchTypes['Image'].id,
+          compliance.images,
+          this.advFilter.imageName,
+          result
+        );
+      } else return false;
+    }
+    if (this.advFilter.selectedDomains.length) {
+      result = this.checkEntity(
+        this.advFilter.matchType4Ns.id,
+        compliance.domains,
+        this.advFilter.selectedDomains.join(','),
+        result
+      );
+    }
+    if (this.advFilter.serviceName) {
+      if (compliance.services && compliance.services.length) {
+        result = this.checkEntity(
+          this.advFilter.matchTypes['Service'].id,
+          compliance.services,
+          this.advFilter.serviceName,
+          result
+        );
+      } else return false;
+    }
+    return result;
+  }
+
+  checkEntity(matchType, entities, pattern, result) {
+    const patterns = pattern.split(',').map(item => item.trim());
+    const theEntity = entities.find(entity => {
+      if (entity && entity.display_name) {
+        if (matchType === 'equal')
+          return patterns.some(item => item === entity.display_name);
+        else return new RegExp(patterns.join('|')).test(entity.display_name);
+      } else {
+        if (matchType === 'equal')
+          return patterns.some(item => item === entity);
+        else return new RegExp(patterns.join('|')).test(entity);
+      }
+    });
+    result = result && !!theEntity;
+    return result;
   }
 
   isAdvFilterOn() {
